# public/openapi.pt-BR.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ZAP CRM BR — API Pública",
    "version": "1.0.0",
    "summary": "API REST para operar uma instância auto-hospedada do ZAP CRM BR a partir dos seus próprios scripts e automações.",
...
    "contact": {
      "name": "ZAP CRM BR",
      "url": "https://github.com/brunopelatieri/zap-crm-br"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [
    {
      "url": "https://{domain}/api/v1",
      "description": "Sua instância auto-hospedada do ZAP CRM BR",
      "variables": {
        "domain": {
          "default": "your-crm.example.com",
          "description": "O domínio onde sua instância do ZAP CRM BR está implantada."
        }
      }
    }
  ],
  "security": [{ "bearerAuth": [] }],
...
  "paths": {
...
    "/broadcasts": {
      "post": {
        "tags": ["Broadcasts"],
        "summary": "Lançar um disparo por template",
        "description": "Lança um disparo por template para uma lista de destinatários. O disparo + suas linhas de destinatário são persistidos imediatamente e os envios saem em segundo plano com um orçamento suave de 60 segundos, então a chamada retorna rápido — consulte `GET /broadcasts/{id}` para acompanhar o progresso. Destinatários são limitados a **1000 por requisição**; números de telefone inválidos são descartados e contados como `rejected`, sem falhar a chamada inteira.\n\nUma requisição próxima do teto de 1000 destinatários pode exceder o orçamento de 60 segundos do envio em segundo plano — ainda não há fila de retentativa durável, então destinatários que ficarem `pending` quando o orçamento se esgota permanecem `pending` em vez de retomar automaticamente. Prefira dividir envios grandes em algumas requisições menores.",
        "operationId": "createBroadcast",
        "security": [{ "bearerAuth": ["broadcasts:send"] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BroadcastCreateRequest"
              },
              "example": {
                "name": "July promo",
                "template_name": "promo_july",
                "template_language": "en_US",
                "recipients": [
                  { "to": "+14155550123", "params": ["Jane"] },
                  { "to": "+14155550124" }
                ]
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "O disparo foi aceito e o envio começou em segundo plano.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/BroadcastCreateResult"
                    }
                  }
                },
                "example": {
                  "data": {
                    "broadcast_id": "8f3c1e2a-0000-4000-8000-000000000020",
                    "status": "sending",
                    "total_recipients": 2,
                    "accepted": 2,
                    "rejected": 0
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "wacrm_live_...",
        "description": "Uma chave de API criada no painel em Configurações → Chaves de API, enviada como `Authorization: Bearer wacrm_live_...`. Os escopos são aplicados por endpoint; veja o escopo exigido de cada operação."
      }
    },
...
    "headers": {
      "RetryAfter": {
        "description": "Segundos até o chamador poder tentar de novo.",
        "schema": { "type": "integer" }
      },
      "RateLimitLimit": {
        "description": "O orçamento de requisições da janela atual.",
        "schema": { "type": "integer" }
      },
      "RateLimitRemaining": {
        "description": "Requisições ainda permitidas na janela atual.",
        "schema": { "type": "integer" }
      },
      "RateLimitReset": {
        "description": "Timestamp Unix (segundos) de quando a janela reinicia.",
        "schema": { "type": "integer" }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Entrada malformada (`bad_request`).",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "error": { "code": "bad_request", "message": "'to' is required" }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Chave de API ausente, malformada, desconhecida, revogada ou expirada (`unauthorized`). Estes casos são deliberadamente indistinguíveis no retorno.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "error": {
                "code": "unauthorized",
                "message": "Missing or invalid API key"
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "A chave é válida mas não tem o escopo que este endpoint exige (`forbidden`).",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "error": {
                "code": "forbidden",
                "message": "This API key is missing the 'messages:send' scope"
              }
            }
          }
        }
      },
...
      "RateLimited": {
        "description": "Limite de taxa por chave excedido — 120 requisições por minuto (`rate_limited`).",
        "headers": {
          "Retry-After": { "$ref": "#/components/headers/RetryAfter" },
          "X-RateLimit-Limit": {
            "$ref": "#/components/headers/RateLimitLimit"
          },
          "X-RateLimit-Remaining": {
            "$ref": "#/components/headers/RateLimitRemaining"
          },
          "X-RateLimit-Reset": { "$ref": "#/components/headers/RateLimitReset" }
        },
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "error": {
                "code": "rate_limited",
                "message": "Rate limit exceeded for this API key"
              }
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "required": ["code", "message"],
        "properties": {
          "code": {
            "type": "string",
            "description": "Código de erro estável e comparável por máquina — trate a lógica sobre este campo, não sobre `message`."
          },
          "message": {
            "type": "string",
            "description": "Texto voltado ao humano, em inglês (ver Idioma das mensagens de erro). Pode ser reformulado; não faça correspondência de padrão sobre ele."
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": { "error": { "$ref": "#/components/schemas/Error" } }
      },
...
      "BroadcastRecipient": {
        "type": "object",
        "required": ["to"],
        "properties": {
          "to": {
            "type": "string",
            "description": "Número de telefone em E.164.",
            "example": "+14155550123"
          },
          "params": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Variáveis posicionais do corpo do template para este destinatário."
          }
        }
      },
      "BroadcastCreateRequest": {
        "type": "object",
        "required": ["template_name", "recipients"],
        "properties": {
          "name": { "type": "string", "description": "Rótulo opcional." },
          "template_name": {
            "type": "string",
            "description": "Um nome de template aprovado pela Meta."
          },
          "template_language": { "type": "string", "default": "en_US" },
          "recipients": {
            "type": "array",
            "minItems": 1,
            "maxItems": 1000,
            "items": { "$ref": "#/components/schemas/BroadcastRecipient" }
          }
        }
      },
      "BroadcastCreateResult": {
        "type": "object",
        "required": [
          "broadcast_id",
          "status",
          "total_recipients",
          "accepted",
          "rejected"
        ],
        "properties": {
          "broadcast_id": { "type": "string", "format": "uuid" },
          "status": { "type": "string", "enum": ["sending"] },
          "total_recipients": { "type": "integer" },
          "accepted": { "type": "integer" },
          "rejected": {
            "type": "integer",
            "description": "Destinatários descartados por número de telefone inválido."
          }
        }
      },
...
    }
  }
}